# apps/iana/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: api-iana
  version: 0.1.0
  description: |
    API HTTP (JSON), só leitura, dos registros de numeração da IANA coletados
    pelo `collector-iana` do BadBlock (tabelas `iana_*`): a quem cada faixa de
    ASN e cada bloco IP (os /8 do IPv4 e os blocos do IPv6 unicast) foi
    entregue (RIR, status, WHOIS, RDAP), os registros de uso especial
    (special-purpose) de IPv4, IPv6 e ASN, que sustentam a regra de **bogon**,
    e o bootstrap RDAP (RFC 9224). Fonte: https://www.iana.org/numbers.

    Versões: os servidores sem `/v1` servem a versão atual (hoje v1); os
    servidores com `/v1` fixam a v1 (`/iana/v1/asn/61610`), com o mesmo
    conteúdo, a mesma chave de cache e o mesmo ETag. Cada rota aparece uma vez
    em `paths`, relativa ao servidor. Saúde (`/health`, `/status`, `/ping`) e
    este manifesto (`/openapi.yaml`) ficam fora do versionamento: sobrescrevem
    `servers` só com os servidores sem `/v1`.

    Todo `GET` também aceita `HEAD`. `OPTIONS` em qualquer caminho responde o
    preflight de CORS (`204`, `Access-Control-Allow-Methods: GET, HEAD,
    OPTIONS`). `/iana` redireciona (`301`) para `/iana/`; qualquer outro
    caminho ou método responde `404` com `code` `not_found`.

    Em todas as respostas: `X-Content-Type-Options: nosniff`,
    `Referrer-Policy: no-referrer`, `Access-Control-Allow-Origin` (padrão
    `*`), `Access-Control-Expose-Headers: ETag, X-Cache, X-Dataset-Version` e
...

...
  license:
    name: MIT
    identifier: MIT
...
servers:
...
paths:
...
  /asn/{asn}:
    get:
      operationId: getASN
      tags: [dados]
      summary: Consulta um ASN
      description: |
        A faixa de `as-numbers-1`/`-2` que contém o ASN (`block`), as faixas
        de `special-purpose-as-numbers` que o contêm (`special`, em ordem de
        início) e a entrada do bootstrap RDAP (`asn.json`) que o contém
        (`rdap`). Nunca responde 404 para um ASN válido: o que a IANA não diz
        sai `null`/`[]` (`block` só é `null` com dados incompletos; faixas
        reservadas e não alocadas não têm `rdap`).
      parameters:
        - $ref: '#/components/parameters/ASN'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: O que a IANA diz sobre o ASN.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPublic'
            X-Cache:
              $ref: '#/components/headers/XCache'
            X-Dataset-Version:
              $ref: '#/components/headers/XDatasetVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ASNResponse'
              examples:
                lacnic:
                  summary: GET /asn/61610
                  value:
                    asn: 61610
                    block:
                      start: 61440
                      end: 61951
                      description: Assigned by LACNIC
                      registry: lacnic
                      whois: whois.lacnic.net
                      rdap_urls: [https://rdap.lacnic.net/rdap/]
                      reference: null
                      registration_date: '2013-06-11'
                    special: []
                    rdap:
                      resource: 61440-61951
                      registry: lacnic
                      urls: [https://rdap.lacnic.net/rdap/]
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
                as_trans:
                  summary: GET /asn/AS23456
                  value:
                    asn: 23456
                    block:
                      start: 23456
                      end: 23456
                      description: AS_TRANS
                      registry: null
                      whois: null
                      rdap_urls: []
                      reference: '[RFC6793]'
                      registration_date: null
                    special:
                      - start: 23456
                        end: 23456
                        reason: AS_TRANS; reserved by [RFC6793]
                        reference: '[RFC6793]'
                    rdap: null
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
...
  /meta:
    get:
      operationId: getMeta
      tags: [meta]
      summary: Estado dos dados
      description: |
        Último dataset aplicado (os 10 arquivos, na ordem fixa do
        collector-iana) e a linha do `collector-iana` em `jobs`. Sem cache e
        sem ETag; responde também antes da primeira carga, com `dataset` e
        `collector` `null`.
      responses:
        '200':
          description: Estado dos dados.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaResponse'
              examples:
                loaded:
                  summary: Com dados (recorte de 2 dos 10 arquivos)
                  value:
                    app: api-iana
                    version: 0.1.0
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
                      sha256: b22ef6b2e76eafd675b5d1a144ad6488720da7c39d56e1d8c09a6cf6401ad06e
                      files:
                        - name: as-numbers-1
                          url: https://www.iana.org/assignments/as-numbers/as-numbers-1.csv
                          sha256: 47f5fe7842029d8077ef749d6b3590293df3252841f1446a188284c5a1839949
                          bytes: 7936
                          rows: 88
                          last_modified: Sat, 19 Sep 2026 00:44:44 GMT
                          publication: null
                        - name: rdap-asn
                          url: https://data.iana.org/rdap/asn.json
                          sha256: 80a0659f933b45130435c0bc7d6143ca2b815b3c5d2ef80c769b985a7821bb65
                          bytes: 4408
                          rows: 159
                          last_modified: Mon, 01 Jun 2026 20:00:01 GMT
                          publication: '2026-06-01T20:00:01Z'
                    collector:
                      app: collector-iana
                      last_sync_at: '2026-09-29T00:49:05Z'
                      last_check_at: '2026-09-29T00:49:05Z'
                      consolidated: false
                empty:
                  summary: Antes da primeira carga
                  value:
                    app: api-iana
                    version: 0.1.0
                    dataset: null
                    collector: null
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
...
  /status:
    servers:
      - url: https://api.badblock.net.br/iana
        description: Produção
      - url: http://127.0.0.1:8107/iana
        description: Desenvolvimento local
    get:
      operationId: getStatus
      tags: [saúde]
      summary: Estado da API
      description: |
        Confere o Postgres, o Valkey e se já há dados. Postgres fora = `error`
        (503); só o Valkey fora = `degraded` (200, a API segue sem cache); sem
        dados ainda = `starting` (200).
      responses:
        '200':
          $ref: '#/components/responses/StatusOK'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/StatusUnavailable'
    post:
      operationId: postStatus
      tags: [saúde]
      summary: Estado da API (POST)
      description: Igual ao `GET /status`; o corpo da requisição é ignorado.
      responses:
        '200':
          $ref: '#/components/responses/StatusOK'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/StatusUnavailable'
...
components:
  parameters:
    ASN:
      name: asn
      in: path
      required: true
      description: Número do AS, de 0 a 4294967295, com ou sem o prefixo `AS` (maiúsculas ou minúsculas).
      schema:
        type: string
        pattern: '^([Aa][Ss])?[0-9]+$'
      example: AS61610
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: |
        ETag de uma resposta anterior (fraco ou forte, numa lista, ou `*`). Se
        bater com o ETag atual, a resposta é `304` sem corpo e sem consultar
        nada.
      schema:
        type: string
      example: W/"3db9c89d25843ce7"
  headers:
    ETag:
      description: |
        ETag fraco, `W/"<hash da versão do dataset + consulta normalizada>"`.
        O mesmo com e sem `/v1`; muda quando a versão dos dados muda.
      schema:
        type: string
      example: W/"3db9c89d25843ce7"
    CacheControlPublic:
      description: Cacheável por 5 minutos.
      schema:
        type: string
        const: public, max-age=300
      example: public, max-age=300
    CacheControlNoStore:
      description: Não cacheável.
      schema:
        type: string
        const: no-store
      example: no-store
    XCache:
      description: |
        `HIT` (veio do Valkey, ou `304`), `MISS` (consultou o Postgres e
        guardou no Valkey) ou `BYPASS` (cache desligado).
      schema:
        type: string
        enum: [HIT, MISS, BYPASS]
      example: MISS
    XDatasetVersion:
      description: Versão dos dados usada na resposta (a mesma de `dataset.version`).
      schema:
        type: string
        format: uuid
      example: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
  responses:
    NotModified:
      description: |
        O `If-None-Match` bateu com o ETag atual: sem corpo e sem consulta ao
        banco (`X-Cache: HIT`).
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Cache-Control:
          $ref: '#/components/headers/CacheControlPublic'
        X-Cache:
          $ref: '#/components/headers/XCache'
        X-Dataset-Version:
          $ref: '#/components/headers/XDatasetVersion'
    BadRequest:
      description: ASN, IP ou tamanho de prefixo inválido (`bad_request`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            asn:
              summary: ASN inválido
              value:
                error:
                  code: bad_request
                  message: 'ASN inválido: use um número de 0 a 4294967295, com ou sem o prefixo AS'
            ip:
              summary: IP inválido
              value:
                error:
                  code: bad_request
                  message: endereço IP inválido
            len:
              summary: Tamanho de prefixo inválido
              value:
                error:
                  code: bad_request
                  message: tamanho de prefixo inválido
...
    DatasetUnavailable:
      description: |
        `dataset_not_ready` (a primeira sincronização do collector-iana ainda
        não terminou) ou `database_unavailable` (falha no Postgres).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            dataset_not_ready:
              summary: Antes da primeira carga
              value:
                error:
                  code: dataset_not_ready
                  message: a primeira sincronização do collector-iana ainda não terminou; tente em alguns minutos
            database_unavailable:
              summary: Falha no Postgres
              value:
                error:
                  code: database_unavailable
                  message: banco de dados indisponível
    DatabaseUnavailable:
      description: Falha no Postgres (`database_unavailable`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: database_unavailable
              message: banco de dados indisponível
    Timeout:
      description: A consulta passou de `DB_TIMEOUT` (`timeout`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: timeout
              message: a consulta demorou demais
    InternalError:
      description: Panic no tratamento da requisição, registrado no log (`internal_error`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: internal_error
              message: erro interno
    StatusOK:
      description: |
        A API responde: `ok` (tudo no ar), `starting` (ainda não há dados) ou
        `degraded` (Valkey fora; a API segue sem cache).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          example:
            success: true
            status: ok
            timestamp: '2026-09-29T01:00:00Z'
            message: api-iana operacional
            checks:
              dataset: ok
              postgres: ok
              valkey: ok
    StatusUnavailable:
      description: Postgres fora (`status` `error`, `success` `false`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
          example:
            success: false
            status: error
            timestamp: '2026-09-29T01:00:00Z'
            message: PostgreSQL indisponível
            checks:
              dataset: ok
              postgres: error
              valkey: ok
  schemas:
    Registry:
      description: |
        RIR responsável: `afrinic`, `apnic`, `arin`, `lacnic`, `ripencc`, ou
        `null` (sem RIR: reservado, IANA, 6to4...; no RDAP, servidor de outro
        operador).
      type: [string, 'null']
      enum: [afrinic, apnic, arin, lacnic, ripencc, null]
    IANADate:
      description: |
        Data como a IANA publica, em texto: `AAAA-MM` ou `AAAA-MM-DD`; `null`
        quando vazia.
      type: [string, 'null']
      pattern: '^[0-9]{4}-[0-9]{2}(-[0-9]{2})?$'
    Dataset:
      type: object
      description: Versão dos dados usada na resposta.
      required: [version, updated_at]
      properties:
        version:
          type: string
          format: uuid
          description: |
            UUID da última carga aplicada pelo collector-iana
            (`iana_run.uuid`); entra na chave de cache e no ETag.
        updated_at:
          type: string
          format: date-time
          description: Quando o collector-iana aplicou essa carga (UTC).
    ASNBlock:
      type: object
      description: Faixa de ASN de `as-numbers-1`/`-2`.
      required: [start, end, description, registry, whois, rdap_urls, reference, registration_date]
      properties:
        start:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: Primeiro ASN da faixa.
        end:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: Último ASN da faixa, inclusive.
        description:
          type: string
          description: Coluna Description como publicada (ex. `Assigned by LACNIC`, `Reserved`).
        registry:
          $ref: '#/components/schemas/Registry'
        whois:
          type: [string, 'null']
          description: Servidor WHOIS da faixa; `null` quando vazio.
        rdap_urls:
          type: array
          description: URLs RDAP da faixa (`[]` se nenhuma).
          items:
            type: string
            format: uri
        reference:
          type: [string, 'null']
          description: Coluna Reference como publicada (ex. `[RFC7607]`); `null` quando vazia.
        registration_date:
          $ref: '#/components/schemas/IANADate'
    SpecialASN:
      type: object
      description: Faixa de ASN de `special-purpose-as-numbers`.
      required: [start, end, reason, reference]
      properties:
        start:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: Primeiro ASN da faixa.
        end:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: Último ASN da faixa, inclusive.
        reason:
          type: string
          description: Coluna Reason for Reservation como publicada.
        reference:
          type: [string, 'null']
          description: Coluna Reference como publicada; `null` quando vazia.
...
    RDAPService:
      type: object
      description: Entrada do bootstrap RDAP (`asn.json`, `ipv4.json` ou `ipv6.json`).
      required: [resource, registry, urls]
      properties:
        resource:
          type: string
          description: Faixa de ASN (`61440-61951`, ou um número só) ou bloco CIDR (`192.0.0.0/8`).
        registry:
          $ref: '#/components/schemas/Registry'
        urls:
          type: array
          description: URLs base do servidor RDAP, na ordem publicada.
          items:
            type: string
            format: uri
    ASNResponse:
      type: object
      required: [asn, block, special, rdap, dataset]
      properties:
        asn:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: ASN consultado, normalizado.
        block:
          description: Faixa que contém o ASN (`null` só com dados incompletos).
          oneOf:
            - $ref: '#/components/schemas/ASNBlock'
            - type: 'null'
        special:
          type: array
          description: Faixas de uso especial que contêm o ASN, em ordem de início (`[]` se nenhuma).
          items:
            $ref: '#/components/schemas/SpecialASN'
        rdap:
          description: Entrada do bootstrap RDAP que contém o ASN, ou `null`.
          oneOf:
            - $ref: '#/components/schemas/RDAPService'
            - type: 'null'
        dataset:
          $ref: '#/components/schemas/Dataset'
...
    MetaResponse:
      type: object
      required: [app, version, dataset, collector]
      properties:
        app:
          type: string
          const: api-iana
          description: Nome do app.
        version:
          type: string
          description: Versão do build da API (`dev` fora de um build com tag).
        dataset:
          description: Último dataset aplicado, ou `null` antes da primeira carga.
          oneOf:
            - $ref: '#/components/schemas/MetaDataset'
            - type: 'null'
        collector:
          description: Linha do `collector-iana` em `jobs`, ou `null` se ainda não existe.
          oneOf:
            - $ref: '#/components/schemas/MetaCollector'
            - type: 'null'
    MetaDataset:
      type: object
      description: Último dataset aplicado pelo collector-iana (`iana_run` com `status = 1`).
      required: [version, updated_at, sha256, files]
      properties:
        version:
          type: string
          format: uuid
          description: Versão dos dados (`iana_run.uuid`).
        updated_at:
          type: string
          format: date-time
          description: Quando a carga foi aplicada (UTC).
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: Hash combinado dos 10 arquivos.
        files:
          type: array
          description: Os 10 arquivos da IANA, na ordem fixa do collector-iana.
          items:
            $ref: '#/components/schemas/MetaFile'
    MetaFile:
      type: object
      description: Um dos arquivos da IANA do dataset aplicado.
      required: [name, url, sha256, bytes, rows, last_modified, publication]
      properties:
        name:
          type: string
          description: |
            Nome do arquivo no collector-iana: `as-numbers-1`, `as-numbers-2`,
            `ipv4-address-space`, `ipv6-unicast-address-assignments`,
            `iana-ipv4-special-registry-1`, `iana-ipv6-special-registry-1`,
            `special-purpose-as-numbers`, `rdap-asn`, `rdap-ipv4` ou
            `rdap-ipv6`.
        url:
          type: string
          format: uri
          description: URL de onde o arquivo foi baixado.
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: SHA-256 do conteúdo.
        bytes:
          type: integer
          format: int64
          description: Tamanho em bytes.
        rows:
          type: [integer, 'null']
          description: Linhas lidas pelo parser; `null` quando o collector não gravou.
        last_modified:
          type: [string, 'null']
          description: Cabeçalho `Last-Modified` da IANA (data HTTP); `null` quando ausente.
        publication:
          type: [string, 'null']
          format: date-time
          description: Campo `publication` do JSON (só nos `rdap-*`); `null` nos demais ou quando ausente.
    MetaCollector:
      type: object
      description: Linha do `collector-iana` na tabela `jobs`.
      required: [app, last_sync_at, last_check_at, consolidated]
      properties:
        app:
          type: string
          const: collector-iana
          description: Nome do coletor.
        last_sync_at:
          type: [string, 'null']
          format: date-time
          description: Última sincronização realizada (UTC).
        last_check_at:
          type: [string, 'null']
          format: date-time
          description: Última verificação dos arquivos na IANA (UTC).
        consolidated:
          type: boolean
          description: Se a consolidação da fase 2 já copiou a carga atual (`jobs.consolidated = 1`).
    Status:
      type: object
      description: Resposta de `/health` e `/status`.
      required: [success, status, timestamp, message, checks]
      properties:
        success:
          type: boolean
          description: '`false` só com `status` `error`.'
        status:
          type: string
          enum: [ok, starting, degraded, error]
          description: |
            `ok` (tudo no ar), `starting` (sem dados ainda), `degraded` (Valkey
            fora) ou `error` (Postgres fora, HTTP 503).
        timestamp:
          type: string
          format: date-time
          description: Hora da verificação (UTC).
        message:
          type: string
          description: Mensagem em PT-BR sobre o estado.
        checks:
          $ref: '#/components/schemas/StatusChecks'
    StatusChecks:
      type: object
      description: Resultado de cada verificação.
      required: [postgres, valkey, dataset]
      properties:
        postgres:
          type: string
          enum: [ok, error]
          description: Ping no Postgres.
        valkey:
          type: string
          enum: [ok, error, disabled]
          description: Ping no Valkey (`disabled` com o cache desligado).
        dataset:
          type: string
          enum: [ok, empty]
          description: '`empty` antes da primeira carga do collector-iana.'
...
    Error:
      type: object
      description: Corpo de toda resposta de erro.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, not_found, dataset_not_ready, database_unavailable, timeout, internal_error]
              description: Código do erro, estável para uso em programas.
            message:
              type: string
              description: Mensagem em PT-BR para pessoas.